feat: merge unregister action to settings page
